refactor[test]: `등수 별로 일치갯수를 알 수 있다` 테스트 위치 조정
