+++ src/test/kotlin/lotto/LottoStatisticsTest.kt
@@ -3,8 +3,11 @@ package lotto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.CsvSource
+import org.junit.jupiter.params.provider.MethodSource
 import java.math.BigDecimal
+import java.util.stream.Stream
 
 class LottoStatisticsTest {
     @Test
@@ -53,5 +56,65 @@ class LottoStatisticsTest {
         val isLoss = statistics.isLoss()
 
         assertThat(isLoss).isFalse()
+    }
+
+
+    @ParameterizedTest
+    @MethodSource("rankTestData")
+    fun `등수 별로 일치갯수를 알 수 있다`(
+        winningNumbers: Lotto,
+        userLottos: Lottos,
+        bonusNumber: Int,
+        expectedRank: Rank,
+    ) {
+        val user = User(Amount(1000))
+        user.buyLotto { userLottos }
+
+        val actual = LottoStatistics.from(user, winningNumbers, LottoNumber(bonusNumber))
+
+        assertThat(actual.machRankCount(expectedRank)).isEqualTo(1)
+    }
+
+    companion object {
+        @JvmStatic
+        fun rankTestData(): Stream<Arguments> =
+            Stream.of(
+                Arguments.of(
+                    Lotto(listOf(1, 2, 3, 4, 5, 6)),
+                    Lottos(listOf(Lotto(listOf(1, 2, 3, 4, 5, 6)))),
+                    7,
+                    Rank.FIRST,
+                ),
+                Arguments.of(
+                    Lotto(listOf(1, 2, 3, 4, 5, 10)),
+                    Lottos(listOf(Lotto(listOf(1, 2, 3, 4, 5, 6)))),
+                    10,
+                    Rank.SECOND,
+                ),
+                Arguments.of(
+                    Lotto(listOf(1, 2, 3, 4, 5, 10)),
+                    Lottos(listOf(Lotto(listOf(1, 2, 3, 4, 5, 20)))),
+                    45,
+                    Rank.THIRD,
+                ),
+                Arguments.of(
+                    Lotto(listOf(1, 2, 3, 4, 11, 13)),
+                    Lottos(listOf(Lotto(listOf(1, 2, 3, 4, 20, 30)))),
+                    5,
+                    Rank.FOURTH,
+                ),
+                Arguments.of(
+                    Lotto(listOf(1, 2, 3, 10, 11, 12)),
+                    Lottos(listOf(Lotto(listOf(1, 2, 3, 4, 5, 6)))),
+                    5,
+                    Rank.FIFTH,
+                ),
+                Arguments.of(
+                    Lotto(listOf(1, 2, 3, 4, 5, 6)),
+                    Lottos(listOf(Lotto(listOf(1, 2, 10, 11, 12, 13)))),
+                    5,
+                    Rank.MISS,
+                ),
+            )
     }
 }
